cleaned up norm errors and added an extra file for handling pipes and fd_duping
