+++ Makefile
@@ -6,7 +6,7 @@
 #    By: plang <plang@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/03/07 14:15:56 by dbarrene          #+#    #+#              #
-#    Updated: 2024/05/13 14:06:18 by plang            ###   ########.fr        #
+#    Updated: 2024/05/14 11:33:22 by dbarrene         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -56,6 +56,7 @@ CSRCS = $(SRCDIR)/parsing.c\
 		$(SRCDIR)/valid_syntax_check.c\
 		$(SRCDIR)/parsing_utils.c\
 		$(SRCDIR)/child_spawning.c\
+		$(SRCDIR)/pipe_handling.c\
 
 OBJS= $(patsubst $(SRCDIR)/%.c, $(OBJDIR)/%.o, $(SRCS))
 BOBJS= $(patsubst $(SRCDIR)/%.c, $(OBJDIR)/%.o, $(BSRCS))
