+++ src/child_spawning.c
@@ -6,56 +6,11 @@
 /*   By: dbarrene <dbarrene@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 13:56:38 by dbarrene          #+#    #+#             */
-/*   Updated: 2024/05/13 14:27:23 by dbarrene         ###   ########.fr       */
+/*   Updated: 2024/05/14 11:36:05 by dbarrene         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-void	open_pipes(t_input *input)
-{	
-	int	i;
-	int	fd;
-
-	i = 0;
-	input->pipes = ft_calloc(input->pipe_count + 1, sizeof (int *));
-	if (!input->pipes)
-	{
-		perror("Minishell malloc fail");
-	}
-	while (i <= input->pipe_count)
-	{
-		input->pipes[i] = malloc(2 * sizeof (int));
-		fd = pipe(input->pipes[i]);
-		if (fd == -1)
-		{
-			ft_printerror("Pipes failed to open\n");
-			// free shit here
-			// exit here
-		}
-		i++;
-	}
-}
-
-void	close_pipes(t_input *input)
-{
-	int	i;
-	int	status;
-
-	i = 0;
-	status = 0;
-	while (input->pipe_count > i)
-	{
-		status = close(input->pipes[i][0]);
-		if (status == -1)
-			ft_printerror("Failed to open first end of pipe at index:%d\n", i);
-		status = close(input->pipes[i][1]);
-		if (status == -1)
-			ft_printerror("Failed to open second end of pipe at index:%d\n", i);
-		free(input->pipes[i++]);
-	}
-}
-
 
 void	wait_for_children(t_input *input)
 {
@@ -76,24 +31,7 @@ void	wait_for_children(t_input *input)
 		}
 		i++;
 	}
-	// probably need to start freeing here;
 }
-
-
-
-
-/* 
- *rewrite the forking function and prep_pids
- Key points that we must consider:
- Relative/Absolute path: Check to see if the binary exists before we even append the path
- in case that the path is provided.
-Permissions: Not only checking to see if the binary for the cmd exists, but also
-that it is executable.
-Builtin: Check to see if the function is a builtin, a flag should be passed to it
-In the case that it is a builtin: It will be run in a child process unless it is
-the only pipeline in the line from readline.
-Splitting the cmd: Keep in mind that the command must be split before even forking since the cmd
-could be a builtin with no pipeline and we still to pass the split cmd
 
 void	prep_child_command(t_args *args)
 {
@@ -104,9 +42,7 @@ void	prep_child_command(t_args *args)
 
 	ep = NULL;
 	i = 0;
-//	printf("Long command: %s\n", args->long_command);
 	args->split_cmds = ft_quotesplit(args->long_command, ' ');
-//	ft_printerror("Command from pcc:%s\n", args->split_cmds[0]);
 	while (args->split_cmds[i])
 	{
 		args->split_cmds[i] = trim_input(args->split_cmds[i], '\"');
@@ -139,12 +75,11 @@ void	prep_child_command(t_args *args)
 	if (split_path)
 		free_2d(split_path);
 }
- *
- * */
+
 void	prep_and_split_command(t_args *args)
 {
 	int	i;
-	
+
 	i = 0;
 	args->envcpy = copy_env(args->envcpy, args->envlist);
 	args->split_cmds = ft_quotesplit(args->long_command, ' ');
@@ -152,6 +87,7 @@ void	prep_and_split_command(t_args *args)
 	while (args->split_cmds[i])
 	{
 		args->split_cmds[i] = trim_input(args->split_cmds[i], '\"');
+		args->split_cmds[i] = trim_input(args->split_cmds[i], '\'');
 		printf("split command at index:%i is :%s\n", i, args->split_cmds[i]);
 		i++;
 	}
@@ -160,43 +96,42 @@ void	prep_and_split_command(t_args *args)
 	if (!args->is_builtin)
 	{
 		if (!access(args->split_cmds[0], F_OK))
+		{
 			if (!access(args->split_cmds[0], X_OK))
 			{
 				args->execpath = ft_strdup(args->split_cmds[0]);
 				return ;
 			}
+		}
 		args->execpath = ft_getenv(args->envlist, "PATH");
-		printf("\nexecpath:%s\n", args->execpath);
 		if (!args->execpath)
 			return ;
 		args->split_path = ft_split(args->execpath, ':');
 		check_path_access(args);
 	}
-//	else if (args->pipecount != 1)
-//		cmd_is_builtin(args->envlist, args->split_cmds);
 }
-
-//		execve(args->execpath, args->split_cmds, args->envcpy);
 
 void	check_path_access(t_args *args)
 {
 	int	i;
 
 	i = 0;
-		while (args->split_path[i])
+	while (args->split_path[i])
+	{
+		args->split_path[i] = ft_strjoin_sep(args->split_path[i],
+				args->split_cmds[0], '/');
+		if (!access(args->split_path[i], F_OK))
 		{
-			args->split_path[i] = ft_strjoin_sep(args->split_path[i], 
-					args->split_cmds[0], '/');
-			if (!access(args->split_path[i], F_OK))
-				if (!access(args->split_path[i], X_OK))
-				{
-					args->execpath = ft_strdup(args->split_path[i]);
-					return ;
-				}
-			if (access(args->split_path[i], F_OK))
-					i++;
+			if (!access(args->split_path[i], X_OK))
+			{
+				args->execpath = ft_strdup(args->split_path[i]);
+				return ;
+			}
 		}
-		free_2d(args->split_path);
+		if (access(args->split_path[i], F_OK))
+			i++;
+	}
+	free_2d(args->split_path);
 }
 
 void	prep_pids(t_input *input)
@@ -205,8 +140,6 @@ void	prep_pids(t_input *input)
 	input->pids = calloc(input->pipe_count + 1, sizeof (pid_t));
 	if (!input->pids)
 		ft_printerror("allocation for PID array failed\n");
-		//free shit and exit or something
-//	printf("Current pipecount in prep_pids:%d\n", input->pipe_count);
 	while (input->pid_index < input->pipe_count)
 	{
 		input->pids[input->pid_index] = fork();
@@ -214,32 +147,26 @@ void	prep_pids(t_input *input)
 		{
 			ft_printerror("Error forking and creating children at index:%d\n", input->pid_index);
 			break ;
-			//should put an error or something;
-			//
 		}
 		if (input->pids[input->pid_index] == 0)
 		{
 			printf("Hello from child at index:%d\n", input->pid_index);
 			if (input->arg_struct[input->pid_index]->redir_count)
 				file_opening(*input->arg_struct[input->pid_index]->redirects);
-//			prep_child_command(input->arg_struct[input->pid_index]);
 			if (input->arg_struct[input->pid_index]->is_builtin)
 			{
-				input->arg_struct[input->pid_index]->builtinstatus = 
-					(cmd_is_builtin(input->arg_struct[input->pid_index]->envlist,
+				input->arg_struct[input->pid_index]->builtinstatus
+					= (cmd_is_builtin(input->arg_struct[input->pid_index]->envlist,
 							input->arg_struct[input->pid_index]->split_cmds));
 				exit (input->arg_struct[input->pid_index]->builtinstatus);
 			}
 			execve(input->arg_struct[input->pid_index]->execpath,
-					input->arg_struct[input->pid_index]->split_cmds,
-					input->arg_struct[input->pid_index]->envcpy);
+				input->arg_struct[input->pid_index]->split_cmds,
+				input->arg_struct[input->pid_index]->envcpy);
 			exit (127);
 			break ;
 		}
-//		printf("Contents of struct:%s\n at index:%d\n", input->arg_struct[input->pid_index]->long_command, input->pid_index);
 		input->pid_index++;
 	}
 	wait_for_children(input);
-	/*forking is somewhat functional, still scared to test more atm but seems to be going well with multiple
-	cmds*/
 }
