+++ src/tokenizing.c
@@ -6,7 +6,7 @@
 /*   By: plang <plang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 15:32:54 by dbarrene          #+#    #+#             */
-/*   Updated: 2024/05/13 14:02:39 by dbarrene         ###   ########.fr       */
+/*   Updated: 2024/05/14 11:37:50 by dbarrene         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,9 +43,9 @@ void	prep_input(char *line, t_input *input)
 		i++;
 	}
 	printf("builtin status:%d\n", input->arg_struct[0]->is_builtin);
-		if (input->pipe_count == 1 && input->arg_struct[0]->is_builtin)
-			return;
-		prep_pids(input);
+	if (input->pipe_count == 1 && input->arg_struct[0]->is_builtin)
+		return ;
+	prep_pids(input);
 }
 
 void	build_struct(t_input *input)
@@ -82,12 +82,10 @@ void	clean_arglist(t_args *args)
 	char	*temp;
 	int		i;
 
-//	printf("address of args struct:%p\n", args);
 	delimset = NULL;
 	i = 0;
 	temp = args->arglist;
 	ft_skip_chars(&temp, ' ');
-//	printf("arglist before pointer arithmetic:%p\n", args->arglist);
 	if (args->redir_count)
 	{
 		if (*temp == '<' || *temp == '>')
@@ -105,14 +103,12 @@ void	clean_arglist(t_args *args)
 		}
 		else
 		{
-//			printf("Temp before going into filename last:%p\n", temp);
 			args->long_command = get_cmd_filename_last(temp);
 			args->tokenized_args = get_file_filename_last(temp);
 		}
 	}
 	else
 		args->long_command = ft_strdup(temp);
-//	printf("long_command:%s\n", args->long_command);
 }
 
 void	tokenize_args(t_args *args)
